+++ tests/unit/test_run_logprep.py
@@ -55,7 +55,7 @@ class TestRunLogprepCli:
                 "logprep.util.configuration.Configuration.as_yaml",
             ),
             (
-                "test dry-run tests/testdata/config/config.yml quickstart/exampledata/input_logdata/test_input.jsonl",
+                "test dry-run tests/testdata/config/config.yml examples/exampledata/input_logdata/test_input.jsonl",
                 "logprep.util.rule_dry_runner.DryRunner.run",
             ),
             (
@@ -255,7 +255,7 @@ class TestRunLogprepCli:
     @mock.patch("logprep.util.rule_dry_runner.DryRunner.run")
     def test_test_dry_run_starts_dry_runner(self, mock_dry_runner):
         config_path = ("tests/testdata/config/config.yml",)
-        events_path = "quickstart/exampledata/input_logdata/test_input.jsonl"
+        events_path = "examples/exampledata/input_logdata/test_input.jsonl"
         result = self.cli_runner.invoke(cli, ["test", "dry-run", *config_path, events_path])
         assert result.exit_code == 0
         mock_dry_runner.assert_called()